Using a package to avoid SBCL error messages.
Created a :format-test package to allow us to shadow the cl:format
function. This allows format to behave normally in an interactive
session and allows format to be redefined with an flet inside the
with-python-test-output macro.

Redefining a built-in while in the cl-user package yields
implementation defined behavior. ELC ignored these cases but SBCL
causes an error.